Remove zsh-incompatible syntax, make custom socket variable explicit.
The original socket PR (#86) allowed the `IMPROMPTU_PORT` environment variable to be overridden with the path to a socket file. The test to check for a socket path used zsh-incompatible syntax `:(-5)` which caused prompt generation to fail.

This makes the use of a new socket explicit with the `IMPROMPTU_UNIX_DOMAIN_SOCKET` environment variable.