Add `impromptu link` to bypass how node links binary files.
While this works well in most scenarios, we need to guarantee that we can run `source impromptu` without anything interfering with our script. Some node version managers shim all global scripts, which can cause broken pipes when we attempt to source the script.

This also eliminates the need to coordinate the global version of impromptu with the local version in $IMPROMPTU_DIR when executing impromptu. Instead, we link the global executable to the local version so we're always running the same code. We may choose to maintain the global version for easy installation.